Fixed incorrect ID datatypes. Fixed many-to-many relationship. Added test controller for people to mess around with.

+++ mylibrary/src/main/java/com/gradprogram/mylibrary/models/BookGenres.java
@@ -13,7 +13,7 @@ import javax.persistence.Id;
 public class BookGenres {
     @Id
     @GeneratedValue
-    private int books_genres_id;
+    private Long books_genres_id;
     private int book_id;
     private int genre_id;
 }
